requirements.txt
google-cloud-aiplatform[agent_engines, adk]
google-adk
cloudpickle
pydantic
pyyaml
requests
ipykernel